Fix default cmscan parameters
Fix a sneaky bug, where the parameters need a hyphen to be passed
correctly to the CMscan analysis constructor rearrange function
Because the params include additional keys with hyphen
(-analysis, -cm_file), there is a 2/6 chance that is will be read
and used silently, and 3/6 chance that the runnable will fail.
Rerunning the runnable makes it more likely to pass this, by chance
alone. But that also means that the other params were not used.

Just readding the hyphens fixes the issue.